Update patch set 1
Patch Set 1:

> > Hi David, was wondering if you might be able to help me get this
 > > plugin building in Bazel. Having the extra directory level
 > between
 > > ‘plugins’ and the actual plugin (example-simpleSshCommand) is
 > > causing issues.
 > 
 > I guess it's entirely unsuported. I noticed Martin started this
 > directory layout during last gerrit hackathon for exmplaes plugin.
 > I'm not convinced yet, that we would like to invest extra efforts
 > for such non standard plugin locations. 

This worked with buck.  If you don't think this layout should be migrated to work with bazel, then how do you think we should organize the example plugin so that it keeps separate example in separate plugins and allows them to be built with bazel?

Luca doing someting similar
 > with owner-plugin, so may be he has a better idea here. Sorry.

Patch-set: 1