Add example how to add a more customized header
If a logo icon should be shown in the header that was not quadratic
or round, styling is not possible in the same way as in the existing
example. E.g. the size is not adjustable in all dimensions.

This change adds another example that shows how to change the title
with some additional freedom, for example allowing rectangular icons.

Change-Id: I80629cfff91232e785758d8c6c1c9fb6edd3688d