[DO NOT MERGE] Use load-time x-plugin support PoC
Adapt the Qualcomm's example by using a more traditional
plugin loader with load-time x-plugin dependency resolution
and Dynamic* mixing.

These two examples are based on Change-Id: I63c9b7f2 and leverage
the introduction of the new Gerrit-ApiModule concept:

Plugin-A. example-depends-on:
  Introduce the generic 'DependencyResolver' API and allocates
  a DynamicMap<DependencyResolver> that any other plugin can use
  to register their implementation.

Plugin-B. example-pd:
  Add a proprietary implementation of a 'DependencyResolver' and
  register it to the published DynamicMap<DepedencyResolver>

The mix-in of the class-loaders of the two plugins happens at load time,
because Plugin-A exposes parts of its bindings as Gerrit-ApiModule which
is understood by the plugin loader.

Plugin-B is built with compile-time dependency on Plugin-A, but it
does not include any parts of its code into his implementation.
At runtime, Plugin-A could be unloaded and Plugin-B will continue to
exist and will fails gracefully if the DynamicMap<DependencyResolver>
has no bindings.

From what I see from this example, the use-cases mentioned in the
design document of Change-Id: I55658f62 are fully satisifed without
having to implement the request-time class-loadign mixin:

- 'depends-on' works also if 'pd' is unloaded: Yes
- 'pd' works also if 'depends-on' is unloaded: Yes, with reduced
  functionality

The only caveat is *IF* 'depends-on' is *never* loaded in the first
place, because in this example the plugin provides both the API
definition and parts of its implementation.

By breaking down the 'depends-on' plugin in two parts, the problem
disappears:

- 'dependson' keeps all the dependency logic implementation, but NOT
   the DynamicItem<DependencyResolver> and the DependencyResolver
   interface.
- 'depends-on-api' has only the DependencyResolver interface and the
  the Gerrit-ApiModule that exposes the DynamicItem<DependencyResolver>.

The above small refactoring would enable also the use-case of 'pd'
loaded when 'depends-on' has ever even existed or loaded.

Checklist against the acceptance criteria of Change-Id: I55658f62

* Can share generic plugins (manifest and depends-on) with the community
  without sharing confidential business logic (PDs plugin): Yes.

* Works with vanilla Gerrit (no libmodules, etc): Yes.

* If we fix a bug in one plugin, we need to only deploy a new version
  of that plugin (unless the fix requires an interface change): Yes,
  with the splitting into 'depends-on' + 'depends-on-api'.

* Load order between plugins does not matter: Yes, with the splitting
  into 'depends-on' + 'depends-on-api', the only Gerrit-ApiModule will
  become 'depends-on-api' whilst 'depends-on' and 'pd' can be
  loaded/unloaded/reloaded in any order.

* Keeping a gerrit server process running 24x7: Yes, there is no restart
  of Gerrit required in any case, even if the 'depends-on-api' would
  change.  Just load/unload/reload of plugins may be required.

* No need for admins or plugin devs to perform any steps beyond regular
  plugin development: Yes, the example is also using standard Bazel
  builds rather than Maven projects with symlinks. Also, the example
  can be built inside Eclipse and debug/run without issues, providing
  the same results you may have in the target runtime environment.

* PDs plugin caches pointers to manifests and other deliverables used
  during resolving Depends-on dependencies: Yes, with the caveat that
  *IF* PDs plugins are directly injecting caches exported by
  'depends-on' outside its interface, then may not function properly
  if those bindings do not exist in the first place, unless they declare
  them as optional Guice injections.

* My IDE can resolve all classes needed? Yes, because all dependencies
  from the API are declared in the BUILD file.

* Does hot plugin swapping without a restart work? Yes.

Change-Id: I6f113602371471f71754f07452abb121da5bd93b