[DO NOT MERGE] Qualcomm example with reflection
In this variant of the Qualcomm's example, the DependencyResolver
interface is moved out of the example-depends-on-api, addressing
Martin's comment on the ability to keep the ability to work without
having even the API interface to invoke the other plugins.

As mentioned in the reply to Martin's comment, this solution would
have to use reflection for the method invocation, because there is
no way to have a static reference to an interface that isn't loaded
into the JVM.

The example-anyapi plugin is a "one-size-fits-all" DynamicMap<Object>
that allows any plugin to expose any interface of any type with any
other plugin. It is scalable because even if you have 1000s of plugins
dynamically loaded/unloaded you'll always need only one 'example-anyapi'
plugin loaded.

The solution is also future-proof because 'example-anyapi' which is the
only plugin required to be loaded first in Gerrit, is never going to
change because it is based on java.lang.Object, the base abstract class
of the entire Java class hierarchy.

With this solution, all remaining issues raised by Qualcomm on the
x-plugins PoC should be fully addressed.

The use of reflection is limited to the example-pd plugin and is not
propagated to Gerrit core or to the plugins' development guidelines,
and it will still allow strong typing when needed.

Change-Id: I862e73c6342120491c99fe463b8e240d0d861f28